requirements.txt
click==8.0.1
configparser==5.0.2
cycler==0.10.0
dataclasses==0.8
Flask==2.0.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
gensim==3.6.0
greenlet==1.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mysqlclient==2.0.3
nltk==3.6.2
numpy==1.19.5
pandas==1.1.5
pickle-mixin==1.0.2
Pillow==8.3.1
pkg-resources==0.0.0
pybind11==2.6.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.11
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
six==1.16.0
smart-open==5.1.0
SQLAlchemy==1.4.20
threadpoolctl==2.1.0
tqdm==4.61.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
